Allow pennies in item value
Although I recoil at using a floating-point value for currency, in this
case I think it's probably OK: we don't perform any calculations, and
it's always going to be a floating-point value in the JSON.

+++ app/models/property.rb
@@ -12,7 +12,7 @@ class Property
   attribute :description, String
   validates :description, presence: true
 
-  attribute :value, Integer
+  attribute :value, Float # Not ideal, but it's floating-point in the JSON too
   validates :value, presence: true
 
   attribute :damage_description, String
